PROTON-2567: Fix fuzz-proactor-receive to work with response file
